Moved import-export preferences to exportSortingTab and renamed the tab
