+++ src/main/java/org/jabref/gui/preferences/FileTab.fxml
@@ -17,34 +17,9 @@
          xmlns="http://javafx.com/javafx/8.0.212" xmlns:fx="http://javafx.com/fxml/1"
          fx:controller="org.jabref.gui.preferences.FileTabView">
     <fx:define>
-        <ToggleGroup fx:id="stringsResolveToggleGroup"/>
         <ToggleGroup fx:id="autolinkToggleGroup"/>
     </fx:define>
     <Label styleClass="titleHeader" text="%File"/>
-
-    <CheckBox fx:id="openLastStartup" text="%Open last edited libraries at startup"/>
-    <HBox alignment="CENTER_LEFT" spacing="10.0">
-        <Label text="%Do not wrap the following fields when saving"/>
-        <TextField fx:id="noWrapFiles" HBox.hgrow="ALWAYS"/>
-    </HBox>
-    <RadioButton fx:id="resolveStringsBibTex" text="%Resolve strings for standard BibTeX fields only"
-                 toggleGroup="$stringsResolveToggleGroup"/>
-    <HBox alignment="CENTER_LEFT" spacing="10.0">
-        <RadioButton fx:id="resolveStringsAll" text="%Resolve strings for all fields except"
-                     toggleGroup="$stringsResolveToggleGroup"/>
-        <TextField fx:id="resolveStringsExcept" HBox.hgrow="ALWAYS"/>
-    </HBox>
-    <HBox alignment="CENTER_LEFT" spacing="10.0">
-        <Label alignment="TOP_LEFT" text="%Newline separator"/>
-        <ComboBox fx:id="newLineSeparator" prefWidth="120.0"/>
-    </HBox>
-    <CheckBox fx:id="alwaysReformatBib" text="%Always reformat BIB file on save and export"/>
-
-    <Label styleClass="sectionHeader" text="%Autosave"/>
-    <HBox alignment="CENTER_LEFT" spacing="10.0">
-        <CheckBox fx:id="autosaveLocalLibraries" text="%Autosave local libraries"/>
-        <Button fx:id="autosaveLocalLibrariesHelp"/>
-    </HBox>
 
     <Label styleClass="sectionHeader" text="%External file links"/>
     <HBox alignment="CENTER_LEFT" spacing="10.0">
