+++ src/main/java/org/jabref/gui/preferences/FileTabView.java
@@ -24,17 +24,6 @@ import de.saxsys.mvvmfx.utils.validation.visualization.ControlsFxVisualizer;
 
 public class FileTabView extends AbstractPreferenceTabView<FileTabViewModel> implements PreferencesTab {
 
-    @FXML private CheckBox openLastStartup;
-    @FXML private TextField noWrapFiles;
-    @FXML private RadioButton resolveStringsBibTex;
-    @FXML private RadioButton resolveStringsAll;
-    @FXML private TextField resolveStringsExcept;
-    @FXML private ComboBox<NewLineSeparator> newLineSeparator;
-    @FXML private CheckBox alwaysReformatBib;
-
-    @FXML private CheckBox autosaveLocalLibraries;
-    @FXML private Button autosaveLocalLibrariesHelp;
-
     @FXML private TextField mainFileDirectory;
     @FXML private CheckBox useBibLocationAsPrimary;
     @FXML private Button autolinkRegexHelp;
@@ -66,21 +55,6 @@ public class FileTabView extends AbstractPreferenceTabView<FileTabViewModel> imp
     public void initialize() {
         this.viewModel = new FileTabViewModel(dialogService, preferences);
 
-        openLastStartup.selectedProperty().bindBidirectional(viewModel.openLastStartupProperty());
-        noWrapFiles.textProperty().bindBidirectional(viewModel.noWrapFilesProperty());
-        resolveStringsBibTex.selectedProperty().bindBidirectional(viewModel.resolveStringsBibTexProperty());
-        resolveStringsAll.selectedProperty().bindBidirectional(viewModel.resolveStringsAllProperty());
-        resolveStringsExcept.textProperty().bindBidirectional(viewModel.resolvStringsExceptProperty());
-        resolveStringsExcept.disableProperty().bind(resolveStringsAll.selectedProperty().not());
-        new ViewModelListCellFactory<NewLineSeparator>()
-                .withText(NewLineSeparator::getDisplayName)
-                .install(newLineSeparator);
-        newLineSeparator.itemsProperty().bind(viewModel.newLineSeparatorListProperty());
-        newLineSeparator.valueProperty().bindBidirectional(viewModel.selectedNewLineSeparatorProperty());
-        alwaysReformatBib.selectedProperty().bindBidirectional(viewModel.alwaysReformatBibProperty());
-
-        autosaveLocalLibraries.selectedProperty().bindBidirectional(viewModel.autosaveLocalLibrariesProperty());
-
         mainFileDirectory.textProperty().bindBidirectional(viewModel.mainFileDirectoryProperty());
         useBibLocationAsPrimary.selectedProperty().bindBidirectional(viewModel.useBibLocationAsPrimaryProperty());
         autolinkFileStartsBibtex.selectedProperty().bindBidirectional(viewModel.autolinkFileStartsBibtexProperty());
@@ -96,7 +70,6 @@ public class FileTabView extends AbstractPreferenceTabView<FileTabViewModel> imp
 
         ActionFactory actionFactory = new ActionFactory(Globals.getKeyPrefs());
         actionFactory.configureIconButton(StandardActions.HELP_REGEX_SEARCH, new HelpAction(HelpFile.REGEX_SEARCH), autolinkRegexHelp);
-        actionFactory.configureIconButton(StandardActions.HELP, new HelpAction(HelpFile.AUTOSAVE), autosaveLocalLibrariesHelp);
 
         validationVisualizer.setDecoration(new IconValidationDecorator());
         Platform.runLater(() -> validationVisualizer.initVisualization(viewModel.mainFileDirValidationStatus(), mainFileDirectory));
