+++ src/main/java/org/jabref/gui/preferences/ImportExportTab.fxml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.VBox?>
+<?import org.jabref.gui.commonfxcontrols.SaveOrderConfigPanel?>
+<?import javafx.scene.control.ToggleGroup?>
+<?import javafx.scene.control.CheckBox?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.RadioButton?>
+<?import javafx.scene.control.ComboBox?>
+<?import javafx.scene.control.Button?>
+
+<fx:root prefWidth="650.0" spacing="10.0" type="VBox" xmlns="http://javafx.com/javafx/11.0.1"
+         xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.jabref.gui.preferences.ImportExportTabView">
+    <Label text="%Import and Export" styleClass="titleHeader"/>
+    <fx:define>
+        <ToggleGroup fx:id="stringsResolveToggleGroup"/>
+    </fx:define>
+
+    <CheckBox fx:id="openLastStartup" text="%Open last edited libraries at startup"/>
+
+    <Label styleClass="sectionHeader" text="%Library file"/>
+    <HBox alignment="CENTER_LEFT" spacing="10.0">
+        <Label text="%Do not wrap the following fields when saving"/>
+        <TextField fx:id="noWrapFiles" HBox.hgrow="ALWAYS"/>
+    </HBox>
+    <RadioButton fx:id="resolveStringsBibTex" text="%Resolve strings for standard BibTeX fields only"
+                 toggleGroup="$stringsResolveToggleGroup"/>
+    <HBox alignment="CENTER_LEFT" spacing="10.0">
+        <RadioButton fx:id="resolveStringsAll" text="%Resolve strings for all fields except"
+                     toggleGroup="$stringsResolveToggleGroup"/>
+        <TextField fx:id="resolveStringsExcept" HBox.hgrow="ALWAYS"/>
+    </HBox>
+    <HBox alignment="CENTER_LEFT" spacing="10.0">
+        <Label alignment="TOP_LEFT" text="%Newline separator"/>
+        <ComboBox fx:id="newLineSeparator" prefWidth="120.0"/>
+    </HBox>
+
+    <Label styleClass="sectionHeader" text="%File reformatting"/>
+    <CheckBox fx:id="alwaysReformatBib" text="%Always reformat BIB file on save and export"/>
+    <SaveOrderConfigPanel fx:id="exportOrderPanel"/>
+
+    <Label styleClass="sectionHeader" text="%Autosave"/>
+    <HBox alignment="CENTER_LEFT" spacing="10.0">
+        <CheckBox fx:id="autosaveLocalLibraries" text="%Autosave local libraries"/>
+        <Button fx:id="autosaveLocalLibrariesHelp"/>
+    </HBox>
+</fx:root>
