+++ src/main/java/org/jabref/gui/preferences/ImportExportTabViewModel.java
@@ -9,14 +9,27 @@ import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleListProperty;
 import javafx.beans.property.SimpleObjectProperty;
+import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.property.StringProperty;
 import javafx.collections.FXCollections;
 
 import org.jabref.model.entry.field.Field;
 import org.jabref.model.entry.field.FieldFactory;
 import org.jabref.model.metadata.SaveOrderConfig;
-import org.jabref.preferences.PreferencesService;
+import org.jabref.preferences.JabRefPreferences;
+import org.jabref.preferences.NewLineSeparator;
 
-public class ExportSortingTabViewModel implements PreferenceTabViewModel {
+public class ImportExportTabViewModel implements PreferenceTabViewModel {
+
+    private final BooleanProperty openLastStartupProperty = new SimpleBooleanProperty();
+    private final StringProperty noWrapFilesProperty = new SimpleStringProperty("");
+    private final BooleanProperty resolveStringsBibTexProperty = new SimpleBooleanProperty();
+    private final BooleanProperty resolveStringsAllProperty = new SimpleBooleanProperty();
+    private final StringProperty resolveStringsExceptProperty = new SimpleStringProperty("");
+    private final ListProperty<NewLineSeparator> newLineSeparatorListProperty = new SimpleListProperty<>();
+    private final ObjectProperty<NewLineSeparator> selectedNewLineSeparatorProperty = new SimpleObjectProperty<>();
+    private final BooleanProperty alwaysReformatBibProperty = new SimpleBooleanProperty();
+
     // SaveOrderConfigPanel
     private final BooleanProperty saveInOriginalProperty = new SimpleBooleanProperty();
     private final BooleanProperty saveInTableOrderProperty = new SimpleBooleanProperty();
@@ -32,14 +45,26 @@ public class ExportSortingTabViewModel implements PreferenceTabViewModel {
     private final ObjectProperty<Field> saveSecondarySortSelectedValueProperty = new SimpleObjectProperty<>(null);
     private final ObjectProperty<Field> saveTertiarySortSelectedValueProperty = new SimpleObjectProperty<>(null);
 
-    private final PreferencesService preferences;
+    private final BooleanProperty autosaveLocalLibraries = new SimpleBooleanProperty();
 
-    ExportSortingTabViewModel(PreferencesService preferences) {
+    private final JabRefPreferences preferences;
+
+    ImportExportTabViewModel(JabRefPreferences preferences) {
         this.preferences = preferences;
     }
 
     @Override
     public void setValues() {
+        openLastStartupProperty.setValue(preferences.getBoolean(JabRefPreferences.OPEN_LAST_EDITED));
+        noWrapFilesProperty.setValue(preferences.get(JabRefPreferences.NON_WRAPPABLE_FIELDS));
+        resolveStringsAllProperty.setValue(preferences.getBoolean(JabRefPreferences.RESOLVE_STRINGS_ALL_FIELDS)); // Flipped around
+        resolveStringsBibTexProperty.setValue(!resolveStringsAllProperty.getValue());
+        resolveStringsExceptProperty.setValue(preferences.get(JabRefPreferences.DO_NOT_RESOLVE_STRINGS_FOR));
+        newLineSeparatorListProperty.setValue(FXCollections.observableArrayList(NewLineSeparator.values()));
+        selectedNewLineSeparatorProperty.setValue(preferences.getNewLineSeparator());
+
+        alwaysReformatBibProperty.setValue(preferences.getBoolean(JabRefPreferences.REFORMAT_FILE_ON_SAVE_AND_EXPORT));
+
         SaveOrderConfig initialExportOrder = preferences.loadExportSaveOrder();
 
         if (initialExportOrder.saveInOriginalOrder()) {
@@ -62,10 +87,23 @@ public class ExportSortingTabViewModel implements PreferenceTabViewModel {
         savePrimaryDescPropertySelected.setValue(initialExportOrder.getSortCriteria().get(0).descending);
         saveSecondaryDescPropertySelected.setValue(initialExportOrder.getSortCriteria().get(1).descending);
         saveTertiaryDescPropertySelected.setValue(initialExportOrder.getSortCriteria().get(2).descending);
+
+        autosaveLocalLibraries.setValue(preferences.getBoolean(JabRefPreferences.LOCAL_AUTO_SAVE));
     }
 
     @Override
     public void storeSettings() {
+
+        preferences.putBoolean(JabRefPreferences.OPEN_LAST_EDITED, openLastStartupProperty.getValue());
+        if (!noWrapFilesProperty.getValue().trim().equals(preferences.get(JabRefPreferences.NON_WRAPPABLE_FIELDS))) {
+            preferences.put(JabRefPreferences.NON_WRAPPABLE_FIELDS, noWrapFilesProperty.getValue());
+        }
+        preferences.putBoolean(JabRefPreferences.RESOLVE_STRINGS_ALL_FIELDS, resolveStringsAllProperty.getValue());
+        preferences.put(JabRefPreferences.DO_NOT_RESOLVE_STRINGS_FOR, resolveStringsExceptProperty.getValue().trim());
+        resolveStringsExceptProperty.setValue(preferences.get(JabRefPreferences.DO_NOT_RESOLVE_STRINGS_FOR));
+        preferences.storeNewLineSeparator(selectedNewLineSeparatorProperty.getValue());
+        preferences.putBoolean(JabRefPreferences.REFORMAT_FILE_ON_SAVE_AND_EXPORT, alwaysReformatBibProperty.getValue());
+
         SaveOrderConfig newSaveOrderConfig = new SaveOrderConfig(
                 saveInOriginalProperty.getValue(),
                 saveInSpecifiedOrderProperty.getValue(),
@@ -79,6 +117,8 @@ public class ExportSortingTabViewModel implements PreferenceTabViewModel {
                         saveTertiarySortSelectedValueProperty.get(),
                         saveTertiaryDescPropertySelected.getValue()));
         preferences.storeExportSaveOrder(newSaveOrderConfig);
+
+        preferences.putBoolean(JabRefPreferences.LOCAL_AUTO_SAVE, autosaveLocalLibraries.getValue());
     }
 
     @Override
@@ -89,6 +129,45 @@ public class ExportSortingTabViewModel implements PreferenceTabViewModel {
     @Override
     public List<String> getRestartWarnings() {
         return null;
+    }
+
+    // General
+
+    public BooleanProperty openLastStartupProperty() {
+        return openLastStartupProperty;
+    }
+
+    public StringProperty noWrapFilesProperty() {
+        return noWrapFilesProperty;
+    }
+
+    public BooleanProperty resolveStringsBibTexProperty() {
+        return resolveStringsBibTexProperty;
+    }
+
+    public BooleanProperty resolveStringsAllProperty() {
+        return resolveStringsAllProperty;
+    }
+
+    public StringProperty resolvStringsExceptProperty() {
+        return resolveStringsExceptProperty;
+    }
+
+    public ListProperty<NewLineSeparator> newLineSeparatorListProperty() {
+        return newLineSeparatorListProperty;
+    }
+
+    public ObjectProperty<NewLineSeparator> selectedNewLineSeparatorProperty() {
+        return selectedNewLineSeparatorProperty;
+    }
+
+    public BooleanProperty alwaysReformatBibProperty() {
+        return alwaysReformatBibProperty;
+    }
+
+    // Autosave
+    public BooleanProperty autosaveLocalLibrariesProperty() {
+        return autosaveLocalLibraries;
     }
 
     // SaveOrderConfigPanel
