--- src/main/java/org/jabref/gui/preferences/ImportTab.fxml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.VBox?>
-<fx:root prefWidth="650.0" spacing="10.0" type="VBox" xmlns="http://javafx.com/javafx/11.0.1"
-         xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.jabref.gui.preferences.ImportTabView">
-
-    <Label styleClass="titleHeader" text="%Import"/>
-
-</fx:root>
