+++ src/main/java/org/jabref/gui/preferences/PreferencesDialogViewModel.java
@@ -52,8 +52,7 @@ public class PreferencesDialogViewModel extends AbstractViewModel {
                 new GroupsTabView(preferences),
                 new EntryEditorTabView(preferences),
                 new CitationKeyPatternTabView(preferences),
-                new ImportTabView(preferences),
-                new ExportSortingTabView(preferences),
+                new ImportExportTabView(preferences),
                 new NameFormatterTabView(preferences),
                 new XmpPrivacyTabView(preferences),
                 new NetworkTabView(preferences),
