+++ src/main/resources/l10n/JabRef_en.properties
@@ -2257,6 +2257,10 @@ Close\ other\ libraries=Close other libraries
 Close\ others=Close others
 Reveal\ in\ file\ explorer=Reveal in file explorer
 
+Import\ and\ Export=Import and Export
+Library\ file=Library file
+File\ reformatting=File reformatting
+
 (\ Note\:\ Press\ return\ to\ commit\ changes\ in\ the\ table\!\ )=( Note\: Press return to commit changes in the table\! )
 
 Reset=Reset
